Allow MINUIT to force invalid parameter values
We are seeing jobs crash where MINUIT is (apparently) setting parameters
to a NaN value.  This shouldn't happen, but it is thought that it is due
to an "INF" likelihood value being handled by MINUIT as a NaN.  This changes
Likelihood::evalFit to "force" the value of a parameter with a log message,
and then try to continue.