Resolve LTS/Issue 615 -- Fix JointProbability
This removes discontinuities in the JointProbability calculation, and
makes sure that the function approaches a zero predicted value
smoothly.  The only implementation had a discontinuity around an
expected MC value of about 1E-16.  The new implementation is finite
for MC > 0.0.  Mathematically, the probability should be infinite at
zero, but is truncated at an expectation of about 1E-308 to guarrantee
the likelihood is continuous.