Resolve Issue 674: Add useful DialInputBuffer error
This fixes the check on the DialInputBuffer input parameters to make
sure that an error message gets printed for the user.  The trap is
usually caused by a mistake in the user input file, so the error
message needs to provide good clues, and throw doesn't help.  Use
sys::exit(EXIT_FAILURE) since this probably isn't caused by a
gundam bug.