Cleanup unit testing build
Previous changes to the test building options were misnamed as
GOOGLE_TEST but disabled all of the CTests.  It also removed more
than just the google tests.  This now separates the removal of
all tests, and also allows removal of just the google tests.
Prints a warning if the tests aren't built since that indicates
a broken build.