By default: enable tests and disable googletest
The correctness of GUNDAM is checked by the regular tests
(ENABLE_TEST=ON), while GoogleTest is doing pretty deep unit tests
that are helpful for figuring the reason for a failure.  Disable the
googletest library by default (SKIP_GOOGLE_TEST=ON) since it depends
on having a modern version available, and are mostly a developer tool.