Make sure that dials are properly updated
To be efficient, the update of the event weights using the
DialInterface infrastructure (i.e. on the CPU) depends on
DialInputBuffer::update() being called exactly once before
reweightEvents(ithread) is used to update the event weights.  This
can't be guarranteed when the GPU is used.  Rather than rework the
logic to use a safer request/acknowledge pattern, this adds special
case control to work around places where DialInputBuffer::update()
will have already been called.  This works for now, but is fragile.