Prefer LogExit to LogThrow for user inputs
LogThrow is great for catching gundam problems, but not as good when
the problem is actually invalid user inputs (the user doesn't care
where gundam caught the problem, and the error forces a stop).
Prefer LogExit when a user problem is found so that GUNDAM goes through
it's exit handlers (and i/o will be flushed).