Make sure samples and dials can't change
Start moving away from the overuse of static methods, and add a lot
more safety to trap changes in the inputs.  The update method is no
longer a static method and should only be called from Build().  The
Build method now takes a reference to the samples and dials, and
makes sure that they are haven't changed (there should never be a
second call).