Start moving away from Cache::Manager statics
This starts reorganizing Cache::Manager to limit the number of static
methods needed to run the Cache::Manager.  The goal is that we will
have "configuration" functions needed to pass command line options and
control (made static since they are used before Cache::Manager
exists).  The Build method, and the Propagate methods.