Add warnings about using statistical PCA
The enablePca field for the fitter engine doesn't do an actual
principle component analysis.  Instead, it independently varies each
parameter by one standard deviation, and eliminates parameters with a
small effect on the statistical LLH.  This is a fairly dangerous
operation for an analysis, so provide a user warning.  If a user wants
to remove a parameter from the fit, it can be explicitly disabled.