Warn user if selected events are not in a bin
Make sure that the likelihood has a bin for all selected events,
and warn the user if there is a mismatch.  The first few unbinned
events will be summarized, and then further output is throttled.  The
number of unbinned events is summarized at the end of the load.  If
GUNDAM_EXIT_ON_UNBINNED_EVENTS is defined, then exit on problems.